requirements.txt
# Core Web Framework
fastapi==0.115.6
uvicorn[standard]

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Database & ORM
sqlalchemy==2.0.21
python-dotenv==1.0.1
pydantic[email]==2.9.0

# Data & NLP
numpy==2.1.3
scikit_learn==1.5.2
sentence_transformers==3.3.1
torch==2.5.1
transformers==4.44.2

# Additional Dependencies
elasticsearch==8.16.0
FlagEmbedding==1.3.3
groq==0.13.0
llama_index==0.12.3
pinecone==5.4.1
pydantic==2.9.0
python_docx==1.1.2
python-multipart==0.0.6